refactor(server): carve out the pull API
Pull-request routing, review state, deferred merge workers, and stack behavior were still coupled to the root server, keeping domain lifecycle and focused tests in the monolithic package.

Give Pull one concrete handler with committed config snapshots and an explicit bounded lifecycle while retaining sync orchestration and genuine ServeHTTP composition boundaries at the root.
