+++ internal/server/httpapi/repository_resolver.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strconv"
 	"strings"
 
 	"go.kenn.io/middleman/internal/db"
@@ -17,6 +18,135 @@ var ErrRepositoryStoreUnavailable = errors.New("repository store unavailable")
 type RepositoryResolver struct {
 	db                   *db.DB
 	providerCapabilities func(platform.Kind, string) (platform.Capabilities, error)
+}
+
+// LookupRoute resolves the provider-aware route tuple through the canonical
+// repository identity path.
+func (r *RepositoryResolver) LookupRoute(
+	ctx context.Context,
+	provider, platformHost, owner, name string,
+) (*db.Repo, error) {
+	owner = strings.Trim(owner, "/ ")
+	name = strings.Trim(name, "/ ")
+	if owner == "" || name == "" {
+		return nil, ErrRepoPathRequired
+	}
+	return r.Lookup(ctx, provider, platformHost, owner+"/"+name)
+}
+
+// RequireRouteCapability combines canonical route lookup with the shared
+// provider-capability fallback policy.
+func (r *RepositoryResolver) RequireRouteCapability(
+	ctx context.Context,
+	provider, platformHost, owner, name, capability string,
+) (*db.Repo, error) {
+	repo, err := r.LookupRoute(ctx, provider, platformHost, owner, name)
+	if err != nil {
+		return nil, ProviderRouteLookupError(err)
+	}
+	if !CapabilityEnabled(r.Ref(*repo).Capabilities, capability) {
+		return nil, UnsupportedCapability(*repo, capability)
+	}
+	return repo, nil
+}
+
+func ProviderRouteLookupError(err error) error {
+	if errors.Is(err, ErrRepoPathRequired) {
+		return BadRequest(CodeBadRequest, err.Error(), nil)
+	}
+	if errors.Is(err, ErrRepoNotFound) {
+		return NotFound(CodeRepoNotFound, "repo not found", nil)
+	}
+	if strings.Contains(err.Error(), "platform_host is required") ||
+		strings.Contains(err.Error(), "unsupported platform") {
+		return BadRequest(CodeBadRequest, err.Error(), nil)
+	}
+	return Internal("get repo failed")
+}
+
+func ProviderKind(repo db.Repo) platform.Kind {
+	if strings.TrimSpace(repo.Platform) == "" {
+		return platform.KindGitHub
+	}
+	return platform.Kind(repo.Platform)
+}
+
+func ProviderHost(repo db.Repo) string {
+	if strings.TrimSpace(repo.PlatformHost) != "" {
+		return repo.PlatformHost
+	}
+	if host, ok := platform.DefaultHost(ProviderKind(repo)); ok {
+		return host
+	}
+	return platform.DefaultGitHubHost
+}
+
+func PlatformRepoRef(repo db.Repo) platform.RepoRef {
+	repoPath := strings.TrimSpace(repo.RepoPath)
+	if repoPath == "" {
+		repoPath = repo.Owner + "/" + repo.Name
+	}
+	numericID, err := strconv.ParseInt(strings.TrimSpace(repo.PlatformRepoID), 10, 64)
+	if err != nil || numericID <= 0 {
+		numericID = 0
+	}
+	return platform.RepoRef{
+		Platform:           ProviderKind(repo),
+		Host:               ProviderHost(repo),
+		Owner:              repo.Owner,
+		Name:               repo.Name,
+		RepoPath:           repoPath,
+		PlatformID:         numericID,
+		PlatformExternalID: repo.PlatformRepoID,
+		WebURL:             repo.WebURL,
+		CloneURL:           repo.CloneURL,
+		DefaultBranch:      repo.DefaultBranch,
+	}
+}
+
+func CapabilityEnabled(caps ProviderCapabilitiesResponse, capability string) bool {
+	switch capability {
+	case "comment_mutation":
+		return caps.CommentMutation
+	case "state_mutation":
+		return caps.StateMutation
+	case "merge_mutation":
+		return caps.MergeMutation
+	case "review_mutation":
+		return caps.ReviewMutation
+	case "workflow_approval":
+		return caps.WorkflowApproval
+	case "ready_for_review":
+		return caps.ReadyForReview
+	case "draft_mutation":
+		return caps.DraftMutation
+	case "issue_mutation":
+		return caps.IssueMutation
+	case "read_labels":
+		return caps.ReadLabels
+	case "label_mutation":
+		return caps.LabelMutation
+	case "assignee_mutation":
+		return caps.AssigneeMutation
+	case "reviewer_mutation":
+		return caps.ReviewerMutation
+	case "thread_reply":
+		return caps.ThreadReply
+	case "thread_resolve":
+		return caps.ThreadResolve
+	case "review_draft_mutation":
+		return caps.ReviewDraftMutation
+	case "review_thread_resolution":
+		return caps.ReviewThreadResolution
+	case "review_suggestion_application":
+		return caps.ReviewSuggestionApplication
+	case "read_review_threads":
+		return caps.ReadReviewThreads
+	case "mutation_head_binding":
+		return caps.MutationHeadBinding
+	default:
+		return false
+	}
 }
 
 type RepositoryResolverDeps struct {
