+++ internal/server/provider_route_wrappers.go
@@ -3,8 +3,19 @@ package server
 import (
 	"context"
 
+	"go.kenn.io/middleman/internal/server/httpapi"
 	"go.kenn.io/middleman/internal/server/workspaceapi"
 )
+
+func repoNumberFromHost(input *repoNumberHostInput) repoNumberInput {
+	return repoNumberInput{
+		Provider:     input.Provider,
+		PlatformHost: input.PlatformHost,
+		Owner:        input.Owner,
+		Name:         input.Name,
+		Number:       input.Number,
+	}
+}
 
 type repoNumberHostInput struct {
 	Provider     string `path:"provider"`
@@ -23,152 +34,6 @@ type resolveItemHostInput struct {
 	ItemType     string `query:"item_type" enum:"pr,issue" doc:"Optional item type hint for providers whose issues and merge requests have separate number spaces."`
 }
 
-type setKanbanStateHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Status string `json:"status"`
-	}
-}
-
-type postCommentHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Body string `json:"body"`
-	}
-}
-
-type editCommentHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	CommentID    int64  `path:"comment_id"`
-	Body         struct {
-		Body string `json:"body"`
-	}
-}
-
-type deleteCommentHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	CommentID    int64  `path:"comment_id"`
-}
-
-type replyToDiscussionHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	DiscussionID string `path:"discussion_id"`
-	Body         struct {
-		Body string `json:"body"`
-	}
-}
-
-type resolveDiscussionHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	DiscussionID string `path:"discussion_id"`
-	Body         struct {
-		Resolved bool `json:"resolved"`
-	}
-}
-
-type createDiffReviewDraftCommentHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Body  string              `json:"body"`
-		Range diffReviewLineRange `json:"range"`
-	}
-}
-
-type editDiffReviewDraftCommentHostInput struct {
-	Provider       string `path:"provider"`
-	PlatformHost   string `path:"platform_host"`
-	Owner          string `path:"owner"`
-	Name           string `path:"name"`
-	Number         int    `path:"number"`
-	DraftCommentID string `path:"draft_comment_id"`
-	Body           struct {
-		Body  string              `json:"body"`
-		Range diffReviewLineRange `json:"range"`
-	}
-}
-
-type deleteDiffReviewDraftCommentHostInput struct {
-	Provider       string `path:"provider"`
-	PlatformHost   string `path:"platform_host"`
-	Owner          string `path:"owner"`
-	Name           string `path:"name"`
-	Number         int    `path:"number"`
-	DraftCommentID string `path:"draft_comment_id"`
-}
-
-type publishDiffReviewDraftHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Body   string `json:"body,omitempty"`
-		Action string `json:"action"`
-	}
-}
-
-type requestChangesPRHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Body            string `json:"body"`
-		ExpectedHeadSHA string `json:"expected_head_sha,omitempty"`
-	}
-}
-
-type applyReviewSuggestionHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		ExpectedHeadSHA string                             `json:"expected_head_sha,omitempty"`
-		Message         string                             `json:"message,omitempty"`
-		Suggestions     []applyReviewSuggestionRequestItem `json:"suggestions"`
-	}
-}
-
-type resolveDiffReviewThreadHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	ThreadID     string `path:"thread_id"`
-}
 type createIssueHostInput struct {
 	Provider     string `path:"provider"`
 	PlatformHost string `path:"platform_host"`
@@ -219,31 +84,13 @@ type getRepoHostInput struct {
 	Name         string `path:"name"`
 }
 
-type setPullLabelsHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         setLabelsRequest
-}
-
 type setIssueLabelsHostInput struct {
 	Provider     string `path:"provider"`
 	PlatformHost string `path:"platform_host"`
 	Owner        string `path:"owner"`
 	Name         string `path:"name"`
 	Number       int    `path:"number"`
-	Body         setLabelsRequest
-}
-
-type setPullAssigneesHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         setAssigneesRequest
+	Body         httpapi.SetLabelsRequest
 }
 
 type setIssueAssigneesHostInput struct {
@@ -252,16 +99,7 @@ type setIssueAssigneesHostInput struct {
 	Owner        string `path:"owner"`
 	Name         string `path:"name"`
 	Number       int    `path:"number"`
-	Body         setAssigneesRequest
-}
-
-type setPullReviewersHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         setReviewersRequest
+	Body         httpapi.SetAssigneesRequest
 }
 
 type commentAutocompleteHostInput struct {
@@ -272,49 +110,6 @@ type commentAutocompleteHostInput struct {
 	Trigger      string `query:"trigger"`
 	Q            string `query:"q"`
 	Limit        int    `query:"limit"`
-}
-
-type approvePRHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Body string `json:"body"`
-		// ExpectedHeadSHA: see approvePRInput.
-		ExpectedHeadSHA string `json:"expected_head_sha,omitempty"`
-	}
-}
-
-type mergePRHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         mergePRInputBody
-}
-
-type deferMergePRHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         mergePRInputBody
-}
-
-type editPRContentHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Body         struct {
-		Title *string `json:"title,omitempty"`
-		Body  *string `json:"body,omitempty"`
-	}
 }
 
 type editIssueContentHostInput struct {
@@ -340,31 +135,6 @@ type githubStateHostInput struct {
 	}
 }
 
-type getDiffHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Whitespace   string `query:"whitespace"`
-	Commit       string `query:"commit" doc:"Scope to a single commit SHA"`
-	From         string `query:"from"   doc:"Start SHA for range diff (inclusive)"`
-	To           string `query:"to"     doc:"End SHA for range diff (inclusive)"`
-}
-
-type getFilePreviewHostInput struct {
-	Provider     string `path:"provider"`
-	PlatformHost string `path:"platform_host"`
-	Owner        string `path:"owner"`
-	Name         string `path:"name"`
-	Number       int    `path:"number"`
-	Path         string `query:"path" doc:"Changed file path to preview"`
-	Side         string `query:"side" enum:"old,new" doc:"Optional diff side to read for context expansion"`
-	Commit       string `query:"commit" doc:"Scope to a single commit SHA"`
-	From         string `query:"from"   doc:"Start SHA for range diff (inclusive)"`
-	To           string `query:"to"     doc:"End SHA for range diff (inclusive)"`
-}
-
 type createIssueWorkspaceHostInput struct {
 	Provider     string `path:"provider"`
 	PlatformHost string `path:"platform_host"`
@@ -375,50 +145,6 @@ type createIssueWorkspaceHostInput struct {
 		GitHeadRef          *string `json:"git_head_ref,omitempty"`
 		ReuseExistingBranch bool    `json:"reuse_existing_branch,omitempty"`
 	}
-}
-
-func repoNumberFromHost(input *repoNumberHostInput) repoNumberInput {
-	return repoNumberInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-	}
-}
-
-func (s *Server) getPullOnHost(ctx context.Context, input *repoNumberHostInput) (*getPullOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.getPull(ctx, &next)
-}
-
-func (s *Server) getMRImportMetadataOnHost(ctx context.Context, input *repoNumberHostInput) (*getMRImportMetadataOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.getMRImportMetadata(ctx, &next)
-}
-
-func (s *Server) setKanbanStateOnHost(ctx context.Context, input *setKanbanStateHostInput) (*statusOnlyOutput, error) {
-	next := setKanbanStateInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.setKanbanState(ctx, &next)
-}
-
-func (s *Server) editPRContentOnHost(ctx context.Context, input *editPRContentHostInput) (*editPRContentOutput, error) {
-	next := editPRContentInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.editPRContent(ctx, &next)
 }
 
 func (s *Server) editIssueContentOnHost(ctx context.Context, input *editIssueContentHostInput) (*editIssueContentOutput, error) {
@@ -433,88 +159,6 @@ func (s *Server) editIssueContentOnHost(ctx context.Context, input *editIssueCon
 	return s.editIssueContent(ctx, &next)
 }
 
-func (s *Server) postCommentOnHost(ctx context.Context, input *postCommentHostInput) (*postCommentOutput, error) {
-	next := postCommentInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.postComment(ctx, &next)
-}
-
-func (s *Server) editCommentOnHost(ctx context.Context, input *editCommentHostInput) (*editCommentOutput, error) {
-	next := editCommentInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		CommentID:    input.CommentID,
-		Body:         input.Body,
-	}
-	return s.editComment(ctx, &next)
-}
-
-func (s *Server) deleteCommentOnHost(ctx context.Context, input *deleteCommentHostInput) (*deleteCommentOutput, error) {
-	return s.deleteComment(ctx, &deleteCommentInput{
-		Provider: input.Provider, PlatformHost: input.PlatformHost,
-		Owner: input.Owner, Name: input.Name, Number: input.Number, CommentID: input.CommentID,
-	})
-}
-
-func (s *Server) replyToDiscussionOnHost(ctx context.Context, input *replyToDiscussionHostInput) (*replyToDiscussionOutput, error) {
-	next := replyToDiscussionInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		DiscussionID: input.DiscussionID,
-		Body:         input.Body,
-	}
-	return s.replyToDiscussion(ctx, &next)
-}
-
-func (s *Server) resolveDiscussionOnHost(ctx context.Context, input *resolveDiscussionHostInput) (*resolveDiscussionOutput, error) {
-	next := resolveDiscussionInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		DiscussionID: input.DiscussionID,
-		Body:         input.Body,
-	}
-	return s.resolveDiscussion(ctx, &next)
-}
-
-func (s *Server) setPullLabelsOnHost(ctx context.Context, input *setPullLabelsHostInput) (*setLabelsOutput, error) {
-	next := setPullLabelsInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.setPullLabels(ctx, &next)
-}
-
-func (s *Server) setPullAssigneesOnHost(ctx context.Context, input *setPullAssigneesHostInput) (*setAssigneesOutput, error) {
-	next := setPullAssigneesInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.setPullAssignees(ctx, &next)
-}
-
 func (s *Server) setIssueAssigneesOnHost(ctx context.Context, input *setIssueAssigneesHostInput) (*setAssigneesOutput, error) {
 	next := setIssueAssigneesInput{
 		Provider:     input.Provider,
@@ -525,140 +169,6 @@ func (s *Server) setIssueAssigneesOnHost(ctx context.Context, input *setIssueAss
 		Body:         input.Body,
 	}
 	return s.setIssueAssignees(ctx, &next)
-}
-
-func (s *Server) setPullReviewersOnHost(ctx context.Context, input *setPullReviewersHostInput) (*setReviewersOutput, error) {
-	next := setPullReviewersInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.setPullReviewers(ctx, &next)
-}
-
-func (s *Server) getDiffReviewDraftOnHost(
-	ctx context.Context,
-	input *repoNumberHostInput,
-) (*getDiffReviewDraftOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.getDiffReviewDraft(ctx, &next)
-}
-
-func (s *Server) createDiffReviewDraftCommentOnHost(
-	ctx context.Context,
-	input *createDiffReviewDraftCommentHostInput,
-) (*createDiffReviewDraftCommentOutput, error) {
-	next := createDiffReviewDraftCommentInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.createDiffReviewDraftComment(ctx, &next)
-}
-
-func (s *Server) editDiffReviewDraftCommentOnHost(
-	ctx context.Context,
-	input *editDiffReviewDraftCommentHostInput,
-) (*editDiffReviewDraftCommentOutput, error) {
-	next := editDiffReviewDraftCommentInput{
-		Provider:       input.Provider,
-		PlatformHost:   input.PlatformHost,
-		Owner:          input.Owner,
-		Name:           input.Name,
-		Number:         input.Number,
-		DraftCommentID: input.DraftCommentID,
-		Body:           input.Body,
-	}
-	return s.editDiffReviewDraftComment(ctx, &next)
-}
-
-func (s *Server) deleteDiffReviewDraftCommentOnHost(
-	ctx context.Context,
-	input *deleteDiffReviewDraftCommentHostInput,
-) (*statusOnlyOutput, error) {
-	next := deleteDiffReviewDraftCommentInput{
-		Provider:       input.Provider,
-		PlatformHost:   input.PlatformHost,
-		Owner:          input.Owner,
-		Name:           input.Name,
-		Number:         input.Number,
-		DraftCommentID: input.DraftCommentID,
-	}
-	return s.deleteDiffReviewDraftComment(ctx, &next)
-}
-
-func (s *Server) publishDiffReviewDraftOnHost(
-	ctx context.Context,
-	input *publishDiffReviewDraftHostInput,
-) (*actionStatusOutput, error) {
-	next := publishDiffReviewDraftInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.publishDiffReviewDraft(ctx, &next)
-}
-
-func (s *Server) discardDiffReviewDraftOnHost(
-	ctx context.Context,
-	input *repoNumberHostInput,
-) (*statusOnlyOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.discardDiffReviewDraft(ctx, &next)
-}
-
-func (s *Server) applyReviewSuggestionsOnHost(
-	ctx context.Context,
-	input *applyReviewSuggestionHostInput,
-) (*applyReviewSuggestionOutput, error) {
-	next := applyReviewSuggestionInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.applyReviewSuggestions(ctx, &next)
-}
-
-func (s *Server) resolveDiffReviewThreadOnHost(
-	ctx context.Context,
-	input *resolveDiffReviewThreadHostInput,
-) (*statusOnlyOutput, error) {
-	next := resolveDiffReviewThreadInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		ThreadID:     input.ThreadID,
-	}
-	return s.resolveDiffReviewThread(ctx, &next)
-}
-
-func (s *Server) unresolveDiffReviewThreadOnHost(
-	ctx context.Context,
-	input *resolveDiffReviewThreadHostInput,
-) (*statusOnlyOutput, error) {
-	next := resolveDiffReviewThreadInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		ThreadID:     input.ThreadID,
-	}
-	return s.unresolveDiffReviewThread(ctx, &next)
 }
 
 func (s *Server) createIssueOnHost(ctx context.Context, input *createIssueHostInput) (*createIssueOutput, error) {
@@ -769,67 +279,6 @@ func (s *Server) getCommentAutocompleteOnHost(ctx context.Context, input *commen
 	return s.getCommentAutocomplete(ctx, &next)
 }
 
-func (s *Server) approvePROnHost(ctx context.Context, input *approvePRHostInput) (*actionStatusOutput, error) {
-	next := approvePRInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.approvePR(ctx, &next)
-}
-
-func (s *Server) requestChangesPROnHost(
-	ctx context.Context,
-	input *requestChangesPRHostInput,
-) (*actionStatusOutput, error) {
-	next := requestChangesPRInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.requestChangesPR(ctx, &next)
-}
-
-func (s *Server) approveWorkflowsOnHost(ctx context.Context, input *repoNumberHostInput) (*actionStatusOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.approveWorkflows(ctx, &next)
-}
-
-func (s *Server) readyForReviewOnHost(ctx context.Context, input *repoNumberHostInput) (*actionStatusOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.readyForReview(ctx, &next)
-}
-
-func (s *Server) mergePROnHost(ctx context.Context, input *mergePRHostInput) (*mergePROutput, error) {
-	next := mergePRInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.mergePR(ctx, &next)
-}
-
-func (s *Server) deferMergePROnHost(ctx context.Context, input *deferMergePRHostInput) (*deferMergePROutput, error) {
-	next := deferMergePRInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.deferMergePR(ctx, &next)
-}
-
 func (s *Server) syncPROnHost(ctx context.Context, input *repoNumberHostInput) (*syncPROutput, error) {
 	next := repoNumberFromHost(input)
 	return s.syncPR(ctx, &next)
@@ -855,18 +304,6 @@ func (s *Server) enqueueIssueSyncOnHost(ctx context.Context, input *repoNumberHo
 	return s.enqueueIssueSync(ctx, &next)
 }
 
-func (s *Server) setPRGitHubStateOnHost(ctx context.Context, input *githubStateHostInput) (*githubStateOutput, error) {
-	next := githubStateInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Body:         input.Body,
-	}
-	return s.setPRGitHubState(ctx, &next)
-}
-
 func (s *Server) setIssueGitHubStateOnHost(ctx context.Context, input *githubStateHostInput) (*githubStateOutput, error) {
 	next := githubStateInput{
 		Provider:     input.Provider,
@@ -877,52 +314,6 @@ func (s *Server) setIssueGitHubStateOnHost(ctx context.Context, input *githubSta
 		Body:         input.Body,
 	}
 	return s.setIssueGitHubState(ctx, &next)
-}
-
-func (s *Server) getCommitsOnHost(ctx context.Context, input *repoNumberHostInput) (*getCommitsOutput, error) {
-	next := repoNumberFromHost(input)
-	return s.getCommits(ctx, &next)
-}
-
-func (s *Server) getDiffOnHost(ctx context.Context, input *getDiffHostInput) (*getDiffOutput, error) {
-	next := getDiffInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Whitespace:   input.Whitespace,
-		Commit:       input.Commit,
-		From:         input.From,
-		To:           input.To,
-	}
-	return s.getDiff(ctx, &next)
-}
-
-func (s *Server) getFilesOnHost(ctx context.Context, input *repoNumberHostInput) (*getFilesOutput, error) {
-	next := getFilesInput(repoNumberFromHost(input))
-	return s.getFiles(ctx, &next)
-}
-
-func (s *Server) getFilePreviewOnHost(ctx context.Context, input *getFilePreviewHostInput) (*getFilePreviewOutput, error) {
-	next := getFilePreviewInput{
-		Provider:     input.Provider,
-		PlatformHost: input.PlatformHost,
-		Owner:        input.Owner,
-		Name:         input.Name,
-		Number:       input.Number,
-		Path:         input.Path,
-		Side:         input.Side,
-		Commit:       input.Commit,
-		From:         input.From,
-		To:           input.To,
-	}
-	return s.getFilePreview(ctx, &next)
-}
-
-func (s *Server) getStackForPROnHost(ctx context.Context, input *repoNumberHostInput) (*getStackForPROutput, error) {
-	next := repoNumberFromHost(input)
-	return s.getStackForPR(ctx, &next)
 }
 
 func (s *Server) createIssueWorkspaceOnHost(ctx context.Context, input *createIssueWorkspaceHostInput) (*workspaceapi.CreateWorkspaceOutput, error) {
