+++ internal/server/pullapi/diff_review_handlers_test.go
@@ -1,4 +1,4 @@
-package server
+package pullapi
 
 import (
 	"strings"
