+++ internal/server/pullapi/helpers.go
@@ -1,0 +1,208 @@
+package pullapi
+
+import (
+	"context"
+	"fmt"
+	"strings"
+	"time"
+
+	"go.kenn.io/middleman/internal/db"
+	"go.kenn.io/middleman/internal/platform"
+	"go.kenn.io/middleman/internal/server/workspaceapi"
+)
+
+type repoNumberPathRef struct {
+	repoID       int64
+	owner        string
+	name         string
+	number       int
+	platformHost string
+}
+
+func buildRepoLookup(repos []db.Repo) map[int64]db.Repo {
+	lookup := make(map[int64]db.Repo, len(repos))
+	for _, repo := range repos {
+		lookup[repo.ID] = repo
+	}
+	return lookup
+}
+
+func workspaceLookupKey(
+	provider, host, owner, name, itemType string,
+	number int,
+) string {
+	if provider == "" {
+		provider = string(platform.KindGitHub)
+	}
+	if host == "" {
+		if defaultHost, ok := platform.DefaultHost(platform.Kind(provider)); ok {
+			host = defaultHost
+		}
+	}
+	return strings.ToLower(provider) + "\x00" +
+		strings.ToLower(host) + "\x00" +
+		strings.ToLower(owner) + "\x00" +
+		strings.ToLower(name) + "\x00" +
+		itemType + "\x00" +
+		fmt.Sprint(number)
+}
+
+func (s *Handler) buildWorkspaceRefLookup(
+	ctx context.Context,
+) (map[string]workspaceapi.WorkspaceRef, error) {
+	workspaces, err := s.db.ListWorkspaces(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("list workspaces: %w", err)
+	}
+	lookup := make(map[string]workspaceapi.WorkspaceRef, len(workspaces))
+	for _, ws := range workspaces {
+		key := workspaceLookupKey(
+			ws.Platform, ws.PlatformHost, ws.RepoOwner, ws.RepoName,
+			ws.ItemType, ws.ItemNumber,
+		)
+		if _, exists := lookup[key]; exists {
+			continue
+		}
+		lookup[key] = workspaceapi.WorkspaceRef{ID: ws.ID, Status: ws.Status}
+	}
+	return lookup, nil
+}
+
+func workspaceRefForRepoItem(
+	lookup map[string]workspaceapi.WorkspaceRef,
+	repo db.Repo,
+	itemType string,
+	number int,
+) *workspaceapi.WorkspaceRef {
+	ref, ok := lookup[workspaceLookupKey(
+		repo.Platform, repoProviderHost(repo), repo.Owner, repo.Name,
+		itemType, number,
+	)]
+	if !ok {
+		return nil
+	}
+	return &ref
+}
+
+func (s *Handler) lookupRepoMap(ctx context.Context) (map[int64]db.Repo, error) {
+	repos, err := s.db.ListRepos(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("list repos: %w", err)
+	}
+	if s.filterRepos != nil {
+		repos = s.filterRepos(repos)
+	}
+	return buildRepoLookup(repos), nil
+}
+
+// filterConfiguredRepos returns only repos that are currently tracked.
+func (s *Handler) lookupMRID(ctx context.Context, ref repoNumberPathRef) (int64, error) {
+	mr, err := s.db.GetMergeRequestByRepoIDAndNumber(
+		ctx, ref.repoID, ref.number,
+	)
+	if err != nil {
+		return 0, err
+	}
+	if mr == nil {
+		return 0, fmt.Errorf(
+			"pull request %s/%s#%d on %s not found",
+			ref.owner, ref.name, ref.number, ref.platformHost,
+		)
+	}
+	return mr.ID, nil
+}
+
+// lookupIssueID resolves the internal issue id from the common route tuple.
+func parseRepoFilter(repo string) (provider, platformHost, owner, name, repoPath string) {
+	repo = strings.Trim(repo, "/ ")
+	if providerPart, hostedPath, ok := strings.Cut(repo, "|"); ok {
+		provider := strings.ToLower(strings.TrimSpace(providerPart))
+		if _, ok := platform.MetadataFor(platform.Kind(provider)); !ok {
+			return "", "", "", "", ""
+		}
+		parts := strings.Split(strings.Trim(hostedPath, "/ "), "/")
+		if len(parts) < 2 {
+			return "", "", "", "", ""
+		}
+		return provider, parts[0], "", "", strings.Join(parts[1:], "/")
+	}
+	return "", "", "", "", ""
+}
+
+func parseRepoFilters(repo string) []db.RepoFilter {
+	parts := strings.Split(repo, ",")
+	filters := make([]db.RepoFilter, 0, len(parts))
+	for _, part := range parts {
+		provider, platformHost, owner, name, repoPath := parseRepoFilter(part)
+		if repoPath != "" {
+			filters = append(filters, db.RepoFilter{
+				Platform:     provider,
+				PlatformHost: platformHost,
+				RepoPath:     repoPath,
+			})
+		} else if owner != "" {
+			filters = append(filters, db.RepoFilter{
+				Platform:     provider,
+				PlatformHost: platformHost,
+				RepoOwner:    owner,
+				RepoName:     name,
+			})
+		}
+	}
+	return filters
+}
+
+func hasInvalidRepoFilter(repo string) bool {
+	for part := range strings.SplitSeq(repo, ",") {
+		part = strings.Trim(part, "/ ")
+		if part == "" {
+			continue
+		}
+		_, _, owner, _, repoPath := parseRepoFilter(part)
+		if owner == "" && repoPath == "" {
+			return true
+		}
+	}
+	return false
+}
+
+func formatUTCRFC3339(t time.Time) string {
+	return t.UTC().Format(time.RFC3339)
+}
+
+func toWorktreeLinkResponses(
+	links []db.WorktreeLink,
+	hostKey string,
+) []worktreeLinkResponse {
+	out := make([]worktreeLinkResponse, len(links))
+	for i, l := range links {
+		out[i] = worktreeLinkResponse{
+			HostKey:        hostKey,
+			WorktreeKey:    l.WorktreeKey,
+			WorktreePath:   l.WorktreePath,
+			WorktreeBranch: l.WorktreeBranch,
+		}
+	}
+	return out
+}
+
+// indexWorktreeLinksByMR groups worktree link responses by
+// merge request ID.
+func indexWorktreeLinksByMR(
+	links []db.WorktreeLink,
+	hostKey string,
+) map[int64][]worktreeLinkResponse {
+	m := make(map[int64][]worktreeLinkResponse)
+	for _, l := range links {
+		m[l.MergeRequestID] = append(
+			m[l.MergeRequestID],
+			worktreeLinkResponse{
+				HostKey:        hostKey,
+				WorktreeKey:    l.WorktreeKey,
+				WorktreePath:   l.WorktreePath,
+				WorktreeBranch: l.WorktreeBranch,
+			},
+		)
+	}
+	return m
+}
