+++ internal/server/pullapi/stack_health_test.go
@@ -1,4 +1,4 @@
-package server
+package pullapi
 
 import (
 	"testing"
